requirements.txt
Django>=5.0,<5.1
psycopg2-binary>=2.9
python-dotenv>=1.0.1
djangorestframework>=3.14.0
